TRELLO-180: refactoring
